requirements.txt
tensorflow==2.2.0
mtcnn==0.1.0
Pillow==7.1.2
pandas==1.0.4